scaffold out admin panel
